Made storage layer pluggable (again)
