+++ core/src/main/java/com/orientechnologies/orient/core/index/OIndexFullText.java
@@ -26,7 +26,7 @@ import com.orientechnologies.orient.core.db.record.OIdentifiable;
 import com.orientechnologies.orient.core.exception.OInvalidIndexEngineIdException;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.serialization.serializer.OStringSerializerHelper;
-import com.orientechnologies.orient.core.storage.impl.local.OAbstractPaginatedStorage;
+import com.orientechnologies.orient.core.storage.OStorage;
 import com.orientechnologies.orient.core.storage.ridbag.sbtree.OIndexRIDContainer;
 
 import java.util.*;
@@ -57,7 +57,7 @@ public class OIndexFullText extends OIndexMultiValues {
 
   private Set<String> stopWords;
 
-  public OIndexFullText(String name, String typeId, String algorithm, int version, OAbstractPaginatedStorage storage,
+  public OIndexFullText(String name, String typeId, String algorithm, int version, OStorage storage,
       String valueContainerAlgorithm, ODocument metadata) {
     super(name, typeId, algorithm, version, storage, valueContainerAlgorithm, metadata);
     acquireExclusiveLock();
