+++ core/src/main/java/com/orientechnologies/orient/core/index/OIndexMultiValues.java
@@ -29,7 +29,7 @@ import com.orientechnologies.orient.core.id.ORID;
 import com.orientechnologies.orient.core.iterator.OEmptyIterator;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.serialization.serializer.stream.OStreamSerializerSBTreeIndexRIDContainer;
-import com.orientechnologies.orient.core.storage.impl.local.OAbstractPaginatedStorage;
+import com.orientechnologies.orient.core.storage.OStorage;
 import com.orientechnologies.orient.core.storage.ridbag.sbtree.OIndexRIDContainer;
 
 import java.util.*;
@@ -41,7 +41,7 @@ import java.util.concurrent.Callable;
  * @author Luca Garulli (l.garulli--(at)--orientdb.com)
  */
 public abstract class OIndexMultiValues extends OIndexAbstract<Set<OIdentifiable>> {
-  public OIndexMultiValues(String name, final String type, String algorithm, int version, OAbstractPaginatedStorage storage,
+  public OIndexMultiValues(String name, final String type, String algorithm, int version, OStorage storage,
       String valueContainerAlgorithm, final ODocument metadata) {
     super(name, type, algorithm, valueContainerAlgorithm, metadata, version, storage);
   }
