+++ core/src/main/java/com/orientechnologies/orient/core/index/OIndexUnique.java
@@ -23,7 +23,7 @@ import com.orientechnologies.orient.core.db.record.OIdentifiable;
 import com.orientechnologies.orient.core.exception.OInvalidIndexEngineIdException;
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.storage.ORecordDuplicatedException;
-import com.orientechnologies.orient.core.storage.impl.local.OAbstractPaginatedStorage;
+import com.orientechnologies.orient.core.storage.OStorage;
 import com.orientechnologies.orient.core.tx.OTransactionIndexChangesPerKey;
 
 /**
@@ -54,7 +54,7 @@ public class OIndexUnique extends OIndexOneValue {
     }
   };
 
-  public OIndexUnique(String name, String typeId, String algorithm, int version, OAbstractPaginatedStorage storage,
+  public OIndexUnique(String name, String typeId, String algorithm, int version, OStorage storage,
       String valueContainerAlgorithm, ODocument metadata) {
     super(name, typeId, algorithm, version, storage, valueContainerAlgorithm, metadata);
   }
