+++ core/src/main/java/com/orientechnologies/orient/core/storage/OCluster.java
@@ -53,7 +53,6 @@ public interface OCluster {
 
   /**
    * Truncates the cluster content. All the entries will be removed.
-   *
    */
   void truncate() throws IOException;
 
@@ -99,9 +98,10 @@ public interface OCluster {
    */
   OPhysicalPosition getPhysicalPosition(OPhysicalPosition iPPosition) throws IOException;
 
-  /** Check if a rid is existent and deleted or not existent
-   *
-   *  return true only if delete flag is set.
+  /**
+   * Check if a rid is existent and deleted or not existent
+   * <p>
+   * return true only if delete flag is set.
    */
   boolean isDeleted(OPhysicalPosition iPPosition) throws IOException;
 
@@ -132,8 +132,6 @@ public interface OCluster {
 
   float recordOverflowGrowFactor();
 
-  boolean isHashBased();
-
   boolean isSystemCluster();
 
   OClusterEntryIterator absoluteIterator();
@@ -148,7 +146,7 @@ public interface OCluster {
 
   /**
    * Hides records content by putting tombstone on the records position but does not delete record itself.
-   *
+   * <p>
    * <p>This method is used in case of record content itself is broken and cannot be read or deleted. So it is emergence method.
    *
    * @param position Position of record in cluster
