+++ core/src/test/java/com/orientechnologies/orient/core/storage/StorageNamingTests.java
@@ -120,6 +120,12 @@ public class StorageNamingTests {
     }
 
     @Override
+    public OStorageOperationResult<Integer> updateRecord(ORecordId rid, boolean updateContent, byte[] content, int version,
+        byte recordType, int mode, ORecordCallback<Integer> callback) {
+      return null;
+    }
+
+    @Override
     public OStorageOperationResult<Integer> recyclePosition(ORecordId iRecordId, byte[] iContent, int iVersion, byte iRecordType) {
       return null;
     }
